# Optimize bc+obj codegen, focus: devel rebuilds.