- Renders bc-file in RAM
- Compares cached bc-file with ram-buffer
- Writes out if changed
- Generates obj if changed
- File writes always go via a temp-file -> rename now
- bc-flags changes incorporated in cached files names
- Removes forking for each render
- Only single threaded concurrency now (could be improved further using core-count threads with one part of the list each)
- On _my_ machine I consistently get about 2s bc-obj-phase in re-compile (when working on the parser in the compiler for instance) compared to about 40s before - (for the bc-obj-phase alone).
  - Full compile in re-compile is about 20s vs ca 60s before (on my slow HDD machine).
- Again: the focus is improving compile time for _re-compile_ in _devel_ mode, when there's obviously changes in just a few files each iteration.
- _Please try and bench it with similar test-case on your machine, comparing with main-compiler_.
- Bonus: I can actually do other things while compiling now - before my machine completely froze during compile.