@asterite 2016-05-13
I think it failed because Travis still uses llvm 3.5... but the compiler doesn't use that version anymore and should work with 3.6. This needs to be fixed somewhere...