@asterite 2016-05-13
It uses llvm 3.5 because we tell it so: https://github.com/crystal-lang/crystal/blob/master/bin/ci#L68