@luislavena 2016-05-13
@asterite the other failure is caused by format changes in `compiler.cr`.