@luislavena 2016-05-13
32bits error might be a fluke of Travis worker memory while OSX error seems to be related to LLVM version Ary mentioned.

@asterite any particular reason we don't build against 3.6 now? What is the status of OSX and LLVM in this matter?