@asterite 2016-05-13
@luislavena Everything should work well with LLVM 3.6. We could even simplify omnibus-crystal to use it, I think right now it uses LLVM 3.5. The reason why it's using an old version is because it works and it's always easy to keep thing unchanged when they work ¯\\\_(ツ)\_/¯