@ozra 2016-05-15
Ok, so I'll let this "hang" meanwhile, since it won't work with <3.something? I'll see what I can do about it, but I'm kinda keen on getting the codegen working against 3.8+, I _might_ look into making that.