@asterite 2016-05-15
@ozra I was planning on testing this PR next week, and updating travis and others if it works well