@asterite 2016-05-16
@ozra Did you try this on linux or mac?