@ozra 2016-05-17
Linux is all I run, are there weird problems? I'll be happy to do what I can to make it work, because it's quite nice with faster compiles :)