@asterite 2016-05-17
I believe the performance you get is similar to the ones here #745 . Basically, forking a lot in linux is slow, but in mac it seems to be fast enough.