@asterite 2016-05-17
I say because I tried your changes on Mac and it always is slower than before.