@ozra 2016-05-17
Aha, interesting. What are the speeds comparatively in Mach?