@asterite 2016-05-19
@ozra I can't remember, but it was like two or three times slower than what we have now.

Probably the best thing to do is to use fork in mac and use spawn in linux, using your PR.