@ozra 2016-05-20
I'll re-work this to use a ifdef, and look in to the "just create n-threads workers handling a list each" and update. I googled around, but found no benches other than pagefaults being slower in Darwin then in Linux kernel, so I haven't figured out what could be the main culprit (and on "what side"). Two different approaches will no doubt be simplest.