@ozra 2016-09-02
@asterite - What's the status on LLVM-versions and such now - I was thinking about re-implementing this with the platform-conditional - the speed up is badly needed for us Linux folks.