@asterite 2016-09-02
@ozra We are still using LLVM 3.5~3.6 for releasing, because LLVM 3.8 needs an entirely different way to be built.