@asterite 2016-12-16
@ozra I finally combined ideas from #745 and your PR and put them all in https://github.com/crystal-lang/crystal/commit/c7921397a6c58e304e6b0dd68ea2746ee604e064 . I used memory buffers from LLVM and compare that to the old .bc before writing them, which is slightly faster than writing to disk. Now also only a limited number of forks (8) is done, so this should be considerably faster on linux. And on mac it's already faster. So thank you for this PR! :-)